requirements.txt
Django
django-etc